cambio tipo float

+++ herencia/Form1.cs
@@ -23,7 +23,7 @@ namespace herencia
             //utilizando la clase hija (Restar)
             Restar resta = new Restar();// creé un nuevo objeto Resta
                                      // va a llamar al metodo de la clase resta Operar 
-            resultado.Text = resta.operar(int.Parse(valor1.Text), int.Parse(valor2.Text)).ToString();
+            resultado.Text = resta.operar(float.Parse(valor1.Text), float.Parse(valor2.Text)).ToString();
         }
 
         private void sumar_Click(object sender, EventArgs e)
@@ -32,7 +32,7 @@ namespace herencia
             //utilizando la clase hija (Sumar)
             Sumar suma = new Sumar();// creé un nuevo objeto SUMA
                                      // va a llamar al metodo de la clase suma Operar 
-            resultado.Text = suma.operar(int.Parse(valor1.Text), int.Parse(valor2.Text)).ToString();
+            resultado.Text = suma.operar(float.Parse(valor1.Text), float.Parse(valor2.Text)).ToString();
         }
 
         private void label1_Click(object sender, EventArgs e)
@@ -51,7 +51,7 @@ namespace herencia
             //utilizando la clase hija (Multiplicar)
             Multiplicar multiplicacion = new Multiplicar();// creé un nuevo objeto multiplicacion
                                         // va a llamar al metodo de la clase multiplicacion Operar 
-            resultado.Text = multiplicacion.operar(int.Parse(valor1.Text), int.Parse(valor2.Text)).ToString();
+            resultado.Text = multiplicacion.operar(float.Parse(valor1.Text), float.Parse(valor2.Text)).ToString();
         }
 
         private void borrar_Click(object sender, EventArgs e)
@@ -70,10 +70,10 @@ namespace herencia
             Dividir division = new Dividir();
 
             // Llamar al método Operar de la clase Dividir
-            double resultadoDivision = division.operar(double.Parse(valor1.Text), double.Parse(valor2.Text));
+            float resultadoDivision = division.operar(float.Parse(valor1.Text), float.Parse(valor2.Text));
 
             // Verificar si el resultado de la división es válido (no es NaN)
-            if (!double.IsNaN(resultadoDivision))
+            if (!float.IsNaN(resultadoDivision))
                 resultado.Text = resultadoDivision.ToString(); // Mostrar el resultado en el cuadro de texto resultado
         }
     }
